+++ src/main/scala/repository/Resolvers.scala
@@ -6,10 +6,10 @@ import repository.schema.DataTypes.Pokemon
 object Resolvers {
   // delegates calculations to the search engine
   class PokemonResolver(data: Vector[Pokemon]) {
-    def pokemonWithHeight(height: Int): Pokemon = Search.pokemonWithHeight(height, data)
-    def pokemonsWithHeight(height: Int): List[Pokemon] = Search.pokemonsWithHeight(height, data).toList
-    def pokemonWithWeight(weight: Int): Pokemon = Search.pokemonWithWeight(weight, data)
-    def pokemonsWithWeight(weight: Int): List[Pokemon] = Search.pokemonsWithWeight(weight, data).toList
+    def pokemonWithHeight(height: Double): Pokemon = Search.pokemonWithHeight(height, data)
+    def pokemonsWithHeight(height: Double): List[Pokemon] = Search.pokemonsWithHeight(height, data).toList
+    def pokemonWithWeight(weight: Double): Pokemon = Search.pokemonWithWeight(weight, data)
+    def pokemonsWithWeight(weight: Double): List[Pokemon] = Search.pokemonsWithWeight(weight, data).toList
     def pokemons: List[Pokemon] = data.toList
   }
